Set self.message default value to "" instead of None.
It is useful when debugging for the CNML synchronizer when
you want to save some time by importing directly interfaces.
You can comment save_nodes() and save_devices() and directly start
by save_interfaces().
With self.message == None, self.message += """string""" fails because
of concatenation of None with a string.